+++ Addons/FactionsValidator/pom.xml
@@ -9,21 +9,23 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>GriefPreventionValidator</artifactId>
-    <name>GriefPreventionValidator</name>
+    <artifactId>FactionsValidator</artifactId>
 
+    <name>FactionsValidator</name>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <maven.compiler.source>1.8</maven.compiler.source>
         <maven.compiler.target>1.8</maven.compiler.target>
     </properties>
+
     <repositories>
         <repository>
-            <id>jitpack.io</id>
-            <url>https://jitpack.io</url>
+            <id>factions</id>
+            <url>https://ci.ender.zone/plugin/repository/everything/</url>
         </repository>
     </repositories>
+
     <dependencies>
         <dependency>
             <groupId>junit</groupId>
@@ -32,10 +34,10 @@
             <scope>test</scope>
         </dependency>
         <dependency>
-            <groupId>com.github.TechFortress</groupId>
-            <artifactId>GriefPrevention</artifactId>
-            <version>16.7.1</version>
+            <groupId>com.massivecraft</groupId>
+            <artifactId>Factions</artifactId>
+            <version>1.6.9.5-U0.5.16</version>
+            <scope>provided</scope>
         </dependency>
     </dependencies>
-    
 </project>
