--- Spigot/src/main/java/me/darkeyedragon/randomtp/validator/RedProtectValidator.java
@@ -1,40 +1,0 @@
-package me.darkeyedragon.randomtp.validator;
-
-/*public class RedProtectValidator implements RandomLocationValidator {
-
-    private final String name;
-    private RedProtect instance;
-    private boolean isLoaded;
-
-    public RedProtectValidator(String name) {
-        this.name = name;
-    }
-
-    @Override
-    public String getName() {
-        return name;
-    }
-
-    @Override
-    public boolean isValid(RandomLocation location) {
-        Location loc = WorldUtil.toLocation(location);
-        return instance.getAPI().getRegion(loc) == null;
-    }
-
-    @Override
-    public boolean isLoaded() {
-        return isLoaded;
-    }
-
-    @Override
-    public void load() {
-        instance = RedProtect.get();
-        this.isLoaded = instance != null;
-    }
-
-    @Override
-    public void setLoaded(boolean loaded) {
-        this.isLoaded = loaded;
-    }
-
-}*/
