--- Spigot/src/main/java/me/darkeyedragon/randomtp/validator/TownyValidator.java
@@ -1,40 +1,0 @@
-package me.darkeyedragon.randomtp.validator;
-
-/*public class TownyValidator implements RandomLocationValidator {
-
-    private final String name;
-    private TownyAPI instance;
-    private boolean isLoaded;
-
-    public TownyValidator(String name) {
-        this.name = name;
-    }
-
-    @Override
-    public String getName() {
-        return name;
-    }
-
-    @Override
-    public boolean isValid(RandomLocation location) {
-        Location loc = WorldUtil.toLocation(location);
-        return instance.isWilderness(loc);
-    }
-
-    @Override
-    public boolean isLoaded() {
-        return isLoaded;
-    }
-
-    @Override
-    public void load() {
-        this.instance = TownyAPI.getInstance();
-        setLoaded(instance != null);
-    }
-
-    @Override
-    public void setLoaded(boolean loaded) {
-        this.isLoaded = loaded;
-    }
-
-}*/
