--- Spigot/src/main/java/me/darkeyedragon/randomtp/validator/Validator.java
@@ -1,28 +1,0 @@
-package me.darkeyedragon.randomtp.validator;
-
-/*public enum Validator {
-    FACTIONS(new FactionsUuidValidator("Factions")),
-    WORLD_GUARD(new WorldGuardValidator("WorldGuard")),
-    GRIEF_PREVENTION(new GriefPreventionValidator("GriefPrevention")),
-    TOWNY(new TownyValidator("Towny")),
-    RED_PROTECT(new RedProtectValidator("RedProtect"));
-
-    final RandomLocationValidator validator;
-
-    Validator(RandomLocationValidator validator) {
-        this.validator = validator;
-    }
-
-    public static RandomLocationValidator getValidator(String name) {
-        for (Validator validator : Validator.values()) {
-            if (validator.validator.getName().equalsIgnoreCase(name)) {
-                return validator.validator;
-            }
-        }
-        return null;
-    }
-
-    public String getName() {
-        return validator.getName();
-    }
-}*/
